update tests based on review comments
